# openapi/harness-core-v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "auto-agent HTTP API",
    "version": "v1",
    "description": "Versioned SaaS HTTP/SSE contract for the routes registered by pkg/server. Tenant and subject ownership are resolved from the authenticated Principal; cross-tenant session access is intentionally reported as 404. This pre-GA v1 contract accepts additive fields and routes, while semantic or breaking changes require a new reviewed contract revision. Operations marked x-harness-contract-first are approved control-plane contracts whose server registration is verified before they can be considered implemented. Optional adapters return the documented 501 error when they are not configured."
  },
  "servers": [{"url": "/"}],
...
  "paths": {
...
    "/v1/sessions/{id}/runs": {
      "get": {"operationId": "listRuns", "tags": ["runs"], "security": [{"BearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/SessionID"}], "responses": {"200": {"description": "durable runs for the session", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunList"}}}}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "501": {"$ref": "#/components/responses/Error"}}},
      "post": {"operationId": "runTurn", "tags": ["runs"], "summary": "Run one synchronous turn as server-sent events", "security": [{"BearerAuth": []}], "x-harness-reconnect": "The SSE stream has no event IDs, Last-Event-ID support, or server-side replay. On disconnect, use GET /v1/sessions/{id}/events with after_seq to retrieve durable events, then inspect GET run state when configured.", "parameters": [{"$ref": "#/components/parameters/SessionID"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunRequest"}}}}, "responses": {"200": {"description": "SSE stream. Event names are persisted session event types; control/store errors use the same Error payload.", "content": {"text/event-stream": {"schema": {"$ref": "#/components/schemas/SSEEvent"}}}}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "409": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}}}
    },
...
    "/v1/sessions/{id}/runs/{runID}/cancel": {"post": {"operationId": "cancelRun", "tags": ["runs"], "security": [{"BearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/SessionID"}, {"$ref": "#/components/parameters/RunID"}], "responses": {"202": {"description": "cancellation requested", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CancelResponse"}}}}, "401": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "409": {"$ref": "#/components/responses/Error"}, "501": {"$ref": "#/components/responses/Error"}}}},
...
    "/v1/resources": {"get": {"operationId": "listResources", "tags": ["resources"], "security": [{"BearerAuth": []}], "responses": {"2XX": {"$ref": "#/components/responses/Success"}, "default": {"$ref": "#/components/responses/Error"}}}},
    "/v1/resources/{key}": {"get": {"operationId": "getResource", "tags": ["resources"], "security": [{"BearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/ResourceKey"}], "responses": {"200": {"description": "resource bytes", "content": {"application/octet-stream": {"schema": {"type": "string", "format": "binary"}}}}, "default": {"$ref": "#/components/responses/Error"}}}, "put": {"operationId": "putResource", "tags": ["resources"], "security": [{"BearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/ResourceKey"}], "requestBody": {"required": true, "content": {"application/octet-stream": {"schema": {"type": "string", "format": "binary"}}}}, "responses": {"2XX": {"$ref": "#/components/responses/Success"}, "default": {"$ref": "#/components/responses/Error"}}}, "delete": {"operationId": "deleteResource", "tags": ["resources"], "security": [{"BearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/ResourceKey"}], "responses": {"2XX": {"$ref": "#/components/responses/Success"}, "default": {"$ref": "#/components/responses/Error"}}}},
...
    "/v1/admin/storage/resources": {"get": {"operationId": "getResourceStorage", "tags": ["admin"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "Storage configuration", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StorageSettingsGetResponse"}}}}, "default": {"$ref": "#/components/responses/Error"}}}, "put": {"operationId": "putResourceStorage", "tags": ["admin"], "security": [{"BearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StorageSettingsPutRequest"}}}}, "responses": {"200": {"description": "Storage configuration saved", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StorageSettingsPutResponse"}}}}, "409": {"description": "Storage revision conflict", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}, "default": {"$ref": "#/components/responses/Error"}}}},
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {"type": "http", "scheme": "bearer", "description": "Built-in account token, or a deployment-injected Authenticator. The resolved Principal supplies tenant, subject, scope, grants, and role; handlers enforce ownership and administrator boundaries. Development header fallback is disabled in production."},
      "RunnerBearer": {"type": "http", "scheme": "bearer", "description": "Private runner token; it cannot access administrative runner task routes."}
    },
    "parameters": {
      "SessionID": {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}},
...
      "RunID": {"name": "runID", "in": "path", "required": true, "schema": {"type": "string"}},
...
      "ResourceKey": {"name": "key", "in": "path", "required": true, "schema": {"type": "string"}, "description": "Slash-delimited resource key."},
...
    },
    "responses": {
      "Success": {"description": "Implemented successful response; exact payload varies by operation."},
      "Error": {"description": "Uniform JSON error envelope", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
    },
    "schemas": {
      "Error": {"type": "object", "required": ["error"], "properties": {"error": {"type": "string"}}, "additionalProperties": false},
...
      "StorageMigrationProgress": {"type": "object", "required": ["state"], "properties": {"state": {"type": "string", "enum": ["syncing", "catching_up", "verifying", "s3_active", "apply_failed", "cancelled"]}, "copied_objects": {"type": "integer", "format": "uint64"}, "copied_bytes": {"type": "integer", "format": "uint64"}, "verified_objects": {"type": "integer", "format": "uint64"}, "verified_bytes": {"type": "integer", "format": "uint64"}, "error_code": {"type": "string"}, "created_at": {"type": "string", "format": "date-time"}, "updated_at": {"type": "string", "format": "date-time"}, "verified_at": {"type": "string", "format": "date-time"}, "activated_at": {"type": "string", "format": "date-time"}}, "additionalProperties": false},
      "StorageSettingsConfiguration": {"type": "object", "required": ["type", "has_access_key", "has_secret"], "properties": {"type": {"type": "string", "enum": ["embedded", "file", "s3"]}, "path": {"type": "string"}, "endpoint": {"type": "string"}, "region": {"type": "string"}, "bucket": {"type": "string"}, "has_access_key": {"type": "boolean"}, "access_key_preview": {"type": "string"}, "path_style": {"type": "boolean"}, "disable_conditional_writes": {"type": "boolean"}, "has_secret": {"type": "boolean"}, "secret_preview": {"type": "string"}, "config_source": {"type": "string", "enum": ["bootstrap", "db", "env_import", "legacy_fallback"]}, "config_status": {"type": "string", "enum": ["active", "inactive", "invalid", "apply_failed", "migration_pending", "restart_pending"]}, "desired_revision": {"type": "string"}}, "additionalProperties": false},
      "StorageSettingsGetResponse": {"type": "object", "required": ["key", "found", "config"], "properties": {"key": {"type": "string"}, "found": {"type": "boolean"}, "config": {"$ref": "#/components/schemas/StorageSettingsConfiguration"}, "active": {"type": "string"}, "desired_type": {"type": "string", "enum": ["embedded", "file", "s3"]}, "active_type": {"type": "string", "enum": ["embedded", "file", "s3"]}, "desired_revision": {"type": "string"}, "active_revision": {"type": "string"}, "restart_pending": {"type": "boolean"}, "error_code": {"type": "string"}, "migration": {"$ref": "#/components/schemas/StorageMigrationProgress"}}, "additionalProperties": false},
      "StorageSettingsPutRequest": {"type": "object", "required": ["type"], "properties": {"type": {"type": "string", "enum": ["embedded", "file", "s3"], "description": "Resources support embedded or s3; sessions support embedded, file, or s3."}, "path": {"type": "string"}, "endpoint": {"type": "string"}, "region": {"type": "string"}, "bucket": {"type": "string"}, "access_key": {"type": "string"}, "secret_key": {"type": "string", "minLength": 1, "writeOnly": true, "description": "Omit to preserve an existing secret; a non-empty value replaces it. An empty value is invalid."}, "clear_secret": {"type": "boolean", "description": "When true, explicitly deletes the stored secret and cannot be sent with secret_key. An S3 configuration without an effective secret is rejected."}, "path_style": {"type": "boolean"}, "disable_conditional_writes": {"type": "boolean", "description": "Session S3 only; resources must reject true."}, "config_status": {"type": "string", "enum": ["active", "inactive", "invalid", "apply_failed", "restart_pending"]}, "expected_revision": {"type": "string", "description": "Compare-and-swap revision. A mismatch returns 409; omission is reserved for an absent row."}}, "additionalProperties": true},
      "StorageSettingsPutResponse": {"type": "object", "required": ["status", "applied"], "properties": {"status": {"const": "saved"}, "applied": {"type": "string"}, "config": {"$ref": "#/components/schemas/StorageSettingsConfiguration"}, "desired_type": {"type": "string", "enum": ["embedded", "file", "s3"]}, "desired_revision": {"type": "string"}, "active_type": {"type": "string", "enum": ["embedded", "file", "s3"]}, "active_revision": {"type": "string"}, "restart_pending": {"type": "boolean"}, "error_code": {"type": "string"}, "migration": {"$ref": "#/components/schemas/StorageMigrationProgress"}}, "additionalProperties": false},
...
      "RunRequest": {"type": "object", "required": ["message"], "properties": {"message": {"type": "string", "minLength": 1}}, "additionalProperties": true},
      "RunRecord": {"type": "object", "required": ["run_id", "session_id", "status"], "properties": {"run_id": {"type": "string"}, "session_id": {"type": "string"}, "status": {"type": "string"}, "attempt": {"type": "integer"}, "generation": {"type": "integer"}}},
      "RunList": {"type": "object", "required": ["runs"], "properties": {"runs": {"type": "array", "items": {"$ref": "#/components/schemas/RunRecord"}}}},
      "CancelResponse": {"type": "object", "required": ["run_id", "status"], "properties": {"run_id": {"type": "string"}, "status": {"type": "string"}}},
      "SSEEvent": {"type": "string", "description": "SSE wire frame: event: <session event type> followed by data: <JSON>. Examples include run/start, assistant/chunk, tool/call, approval/requested, run/end, store/error, and control/error."},
...
    }
  }
}